Get coordinates from manual address
This is a rough approach: if a manual address is entered, we loop through all establishments and we keep the first that returns true to this statement: the entered address includes the nombre_ciudad of the establishment
This can result in inaccurate results (that's why we zoom out a bit). If no match is found, we show default coordinates: the geographic center of the country (in which case we zoom out a lot).